feat: refactor key-auth identity realms field and update plugin configuration
