+++ packages/entities/entities-plugins/src/components/free-form/plugins/jwt-signer.ts
@@ -1,3 +1,3 @@
 import { definePluginConfig } from '../shared/define-plugin-config'
 
-export default definePluginConfig({})
+export default definePluginConfig()
