+++ packages/entities/entities-plugins/src/components/free-form/shared/define-plugin-config.ts
@@ -1,7 +1,7 @@
 import CommonForm from '../Common/CommonForm.vue'
 import type { PluginFormConfig } from './types'
 
-export function definePluginConfig(config: Partial<PluginFormConfig>): PluginFormConfig {
+export function definePluginConfig(config: Partial<PluginFormConfig> = {}): PluginFormConfig {
   return {
     ...config,
     component: config.component ?? CommonForm,
